+++ cookbooks/wt_data_deleter/recipes/default.rb
@@ -1,28 +1,26 @@
 #
-# Cookbook Name:: wt_datadeleter
+# Cookbook Name:: wt_data_deleter
 # Recipe:: default
 # Author:: Kendrick Martin
 #
-# Copyright 2012, Webtrends, Inc
+# Copyright 2012, Webtrends Inc.
 #
 # All rights reserved - Do Not Redistribute
 #
-# This recipe installs the needed components to full setup/configure the Search service
-#
+
 if ENV["deploy_build"] == "true" then
   log "The deploy_build value is true so un-deploying first"
-  include_recipe "wt_datadeleter::uninstall"
+  include_recipe "wt_data_deleter::uninstall"
 else
   log "The deploy_build value is not set or is false so we will only update the configuration"
 end
 
 # get parameters
-download_url = node['wt_datadeleter']['download_url']
-master_host = node['wt_masterdb']['master_host']
+download_url = node['wt_data_deleter']['download_url']
 
 # destinations
-install_dir = File.join(node['wt_common']['install_dir_windows'], node['wt_datadeleter']['install_dir'].gsub(/[\\\/]+/,"\\"))
-log_dir = File.join(node['wt_common']['install_dir_windows'], node['wt_datadeleter']['log_dir'].gsub(/[\\\/]+/,"\\"))
+install_dir = File.join(node['wt_common']['install_dir_windows'], node['wt_data_deleter']['install_dir'].gsub(/[\\\/]+/,"\\"))
+log_dir = File.join(node['wt_common']['install_dir_windows'], node['wt_data_deleter']['log_dir'].gsub(/[\\\/]+/,"\\"))
 
 # get data bag items
 auth_data = data_bag_item('authorization', node.chef_environment)
@@ -48,12 +46,6 @@ wt_base_icacls node['wt_common']['install_dir_windows'] do
   perm :modify
 end
 
-wt_base_netlocalgroup "Performance Monitor Users" do
-  user svcuser
-  returns [0, 2]
-  action :add
-end
-
 if ENV["deploy_build"] == "true" then
 
   # unzip the install package
@@ -65,7 +57,7 @@ if ENV["deploy_build"] == "true" then
   template "#{install_dir}\\DataDeleter.exe.config" do
     source "DataDeleter.erb"
     variables(
-      :master_host => master_host,
+      :master_host => node['wt_masterdb']['master_host'],
       :hbase_location => node['hbase']['location'],
       :hbase_dc_id => node['wt_analytics_ui']['fb_data_center_id'],
       :hbase_pod_id => node['wt_common']['pod_id'],
@@ -79,7 +71,7 @@ if ENV["deploy_build"] == "true" then
   template "#{install_dir}\\DeletionScheduler.exe.config" do
     source "DeletionScheduler.erb"
     variables(
-      :master_host => master_host,
+      :master_host => node['wt_masterdb']['master_host'],
       :hbase_location => node['hbase']['location'],
       :hbase_dc_id => node['wt_analytics_ui']['fb_data_center_id'],
       :hbase_pod_id => node['wt_common']['pod_id'],
@@ -87,12 +79,11 @@ if ENV["deploy_build"] == "true" then
       :cass_thrift_port => node['cassandra']['cassandra_thrift_port'],
       :report_column => node['cassandra']['cassandra_report_column'],
       :metadata_column => node['cassandra']['cassandra_meta_column']
-
     )
   end
 
   powershell "install data deleter" do
-    environment({'install_dir' => install_dir, 'service_binary' => node['wt_datadeleter']['datadeleter_binary']})
+    environment({'install_dir' => install_dir, 'service_binary' => node['wt_data_deleter']['datadeleter_binary']})
     code <<-EOH
     $binary_path = $env:install_dir + "\\" + $env:service_binary
           &$binary_path --install
@@ -100,12 +91,13 @@ if ENV["deploy_build"] == "true" then
   end
 
   powershell "install deletion scheduler" do
-    environment({'install_dir' => install_dir, 'service_binary' => node['wt_datadeleter']['deletionscheduler_binary']})
+    environment({'install_dir' => install_dir, 'service_binary' => node['wt_data_deleter']['deletionscheduler_binary']})
     code <<-EOH
     $binary_path = $env:install_dir + "\\" + $env:service_binary
           &$binary_path --install
-
     EOH
   end
+
   share_wrs
+
 end