+++ cookbooks/wt_platformscheduler/recipes/agent_uninstall.rb
@@ -3,13 +3,14 @@
 # Recipe:: agent_uninstall
 # Author:: Kendrick Martin
 #
-# Copyright 2012, Webtrends
+# Copyright 2012, Webtrends Inc.
 #
 # All rights reserved - Do Not Redistribute
-# This recipe uninstalls scheduler agent
+#
+# This recipe uninstalls scheduler agent.
+#
 
-
-windows_package "Webtrends VDM Scheduler Agent" do
+windows_package "WebtrendsVDMSchedulerAgent" do
 	action :remove
 	ignore_failure true
 end